cast to uint32_t in order to fix ambiguity for overloaded methods (msvc)
svn path=/trunk/KDE/kdebase/runtime/; revision=768816

+++ trashthroughanalyzer.cpp
@@ -96,7 +96,7 @@ InputStream* TrashThroughAnalyzer::connectInputStream(InputStream* in) {
     idx->addValue(factory->originalpathField,
         (const char*)trashInfo.origPath.toUtf8());
     idx->addValue(factory->dateofdeletionField,
-        trashInfo.deletionDate.toTime_t());
+        (uint32_t)trashInfo.deletionDate.toTime_t());
     return in;
 }
